Dark mode and preview enhancements
