+++ compose/src/main/java/com/bottlerocketstudios/compose/auth/AuthCodeScreen.kt
@@ -4,9 +4,10 @@ import android.annotation.SuppressLint
 import android.net.Uri
 import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
@@ -19,11 +20,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import com.bottlerocketstudios.compose.R
+import com.bottlerocketstudios.compose.resources.Colors
 import com.bottlerocketstudios.compose.resources.Dimens
 import com.bottlerocketstudios.compose.util.Preview
-import com.bottlerocketstudios.compose.util.PreviewAllDevices
+import com.bottlerocketstudios.compose.util.PreviewAll
 import com.bottlerocketstudios.compose.util.asMutableState
 import com.bottlerocketstudios.compose.widgets.OutlinedSurfaceButton
 import com.bottlerocketstudios.compose.widgets.PrimaryButton
@@ -47,24 +48,6 @@ fun AuthCodeScreen(state: AuthCodeState, navigator: WebViewNavigator) {
     }
 }
 
-@PreviewAllDevices
-@Composable
-fun AuthCodePreview() {
-    Preview {
-        AuthCodeScreen(
-            state = AuthCodeState(
-                requestUrl = "".asMutableState(),
-                devOptionsEnabled = true,
-                onAuthCode = {},
-                onLoginClicked = {},
-                onSignupClicked = {},
-                onDevOptionsClicked = {},
-                showToolbar = {}
-            ),
-            navigator = rememberWebViewNavigator()
-        )
-    }
-}
 
 // FIXME: javaScriptEnabled = true can result in removal from Play Store
 // https://support.google.com/faqs/answer/7668153?hl=en
@@ -90,7 +73,9 @@ fun RequestAuth(url: String, onAuthCode: (String) -> Unit, navigator: WebViewNav
 private fun AuthCodeContent(state: AuthCodeState) {
     Column(
         modifier = Modifier
-            .fillMaxHeight()
+            // .fillMaxHeight()
+            .fillMaxSize()
+            .background(Colors.background)
             .verticalScroll(rememberScrollState()),
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.Center,
@@ -107,6 +92,7 @@ private fun AuthCodeContent(state: AuthCodeState) {
             text = stringResource(R.string.auth_code_subtitle),
             textAlign = TextAlign.Center,
             style = MaterialTheme.typography.h2,
+            color = Colors.onBackground,
         )
 
         Text(
@@ -115,6 +101,7 @@ private fun AuthCodeContent(state: AuthCodeState) {
                 .padding(start = Dimens.grid_2, end = Dimens.grid_2),
             text = stringResource(R.string.auth_code_description),
             style = MaterialTheme.typography.h3,
+            color = Colors.onBackground,
             textAlign = TextAlign.Center,
         )
 
@@ -150,10 +137,11 @@ private fun AuthCodeContent(state: AuthCodeState) {
     }
 }
 
-@Preview(showSystemUi = true)
+
+@PreviewAll
 @Composable
-fun AuthCodeDarkPreview() {
-    Preview(darkTheme = true) {
+fun AuthCodePreview() {
+    Preview {
         AuthCodeScreen(
             state = AuthCodeState(
                 requestUrl = "".asMutableState(),
@@ -168,3 +156,4 @@ fun AuthCodeDarkPreview() {
         )
     }
 }
+
