+++ compose/src/main/java/com/bottlerocketstudios/compose/snippets/CreateSnippetScreen.kt
@@ -1,6 +1,5 @@
 package com.bottlerocketstudios.compose.snippets
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -16,7 +15,6 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -25,13 +23,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.bottlerocketstudios.compose.R
 import com.bottlerocketstudios.compose.resources.ArchitectureDemoTheme
-import com.bottlerocketstudios.compose.resources.Colors
 import com.bottlerocketstudios.compose.resources.Dimens
 import com.bottlerocketstudios.compose.util.Preview
+import com.bottlerocketstudios.compose.util.PreviewAll
 import com.bottlerocketstudios.compose.util.asMutableState
 import com.bottlerocketstudios.compose.widgets.OutlinedInputField
 import com.bottlerocketstudios.launchpad.compose.bold
@@ -44,7 +41,6 @@ fun CreateSnippetScreen(state: CreateSnippetScreenState) {
     Column(
         modifier = Modifier
             .fillMaxSize()
-            .background(Colors.background)
     ) {
         OutlinedInputField(
             text = uiState.title.value,
@@ -148,7 +144,7 @@ fun LabelledCheckbox(isPrivate: Boolean, onCheckedChange: (Boolean) -> Unit) {
     }
 }
 
-@Preview(showSystemUi = true)
+@PreviewAll
 @Composable
 fun CreateSnippetScreenPreview() {
     Preview {
@@ -158,15 +154,6 @@ fun CreateSnippetScreenPreview() {
     }
 }
 
-@Preview(showSystemUi = true)
-@Composable
-fun CreateSnippetScreenDarkPreview() {
-    Preview(darkTheme = true) {
-        CreateSnippetScreen(
-            state = previewState()
-        )
-    }
-}
 
 @Composable
 fun previewState() = CreateSnippetScreenState(
@@ -184,16 +171,3 @@ fun previewState() = CreateSnippetScreenState(
 )
 
 
-data class CreateSnippetScreenState(
-    val title: State<String>,
-    val filename: State<String>,
-    val contents: State<String>,
-    val isPrivate: State<Boolean>,
-    val creationFailed: State<Boolean>,
-    val createEnabled: State<Boolean>,
-    val onTitleChanged: (String) -> Unit,
-    val onFilenameChanged: (String) -> Unit,
-    val onContentsChanged: (String) -> Unit,
-    val onPrivateChanged: (Boolean) -> Unit,
-    val onCreateClicked: () -> Unit,
-)
