+++ compose/src/main/java/com/bottlerocketstudios/compose/snippets/SnippetsBrowserScreenState.kt
@@ -1,0 +1,9 @@
+package com.bottlerocketstudios.compose.snippets
+
+import androidx.compose.runtime.State
+
+data class SnippetsBrowserScreenState(
+    val snippets: State<List<SnippetUiModel>>,
+    val createVisible: State<Boolean>,
+    val onCreateSnippetClicked: () -> Unit
+)
