--- compose/src/main/java/com/bottlerocketstudios/compose/snippets/SnippetsMockData.kt
@@ -1,13 +1,0 @@
-package com.bottlerocketstudios.compose.snippets
-
-import com.bottlerocketstudios.compose.util.StringIdHelper
-
-internal val mockSnippet1 = SnippetUiModel(
-    id = "123456789XYZ",
-    workspaceSlug = "workspace_slug",
-    title = "Private",
-    userName = "Bob Ross",
-    formattedLastUpdatedTime = StringIdHelper.Raw(""),
-)
-
-internal val listOfMockSnippets = listOf(mockSnippet1, mockSnippet1)
