+++ compose/src/main/java/com/bottlerocketstudios/compose/util/PreviewUtil.kt
@@ -1,6 +1,8 @@
 package com.bottlerocketstudios.compose.util
 
 import android.annotation.SuppressLint
+import android.content.res.Configuration.UI_MODE_NIGHT_YES
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.tooling.preview.Devices
@@ -9,7 +11,7 @@ import com.bottlerocketstudios.compose.resources.ArchitectureDemoTheme
 
 @Composable
 fun Preview(
-    darkTheme: Boolean = false,
+    darkTheme: Boolean = isSystemInDarkTheme(),
     content: @Composable () -> Unit
 ) = ArchitectureDemoTheme(
     darkTheme = darkTheme,
@@ -17,10 +19,18 @@ fun Preview(
 )
 
 @SuppressLint("UnrememberedMutableState")
-@Composable
 fun <T> T.asMutableState() = mutableStateOf(this)
 
-@Preview(showSystemUi = true, device = Devices.FOLDABLE)
-@Preview(showSystemUi = true, device = Devices.TABLET)
-@Preview(showSystemUi = true, device = Devices.DEFAULT)
-annotation class PreviewAllDevices
+@Preview(showSystemUi = true, device = Devices.FOLDABLE, group = "Foldable")
+@Preview(showSystemUi = true, device = Devices.TABLET, group = "Tablet")
+@PreviewLightDark
+annotation class PreviewAll
+
+@Preview(showSystemUi = true, device = Devices.DEFAULT, group = "Default")
+@Preview(showSystemUi = true, device = Devices.DEFAULT, group = "Default", uiMode = UI_MODE_NIGHT_YES)
+annotation class PreviewLightDark
+
+@Preview(showBackground = true)
+@Preview(showBackground = true, uiMode = UI_MODE_NIGHT_YES)
+annotation class PreviewComposable
+
